Funcionalidad Completa

+++ src/main/java/co/edu/escuelaing/arem/designprimer/MyLinkedList.java
@@ -21,7 +21,7 @@ class MyLinkedList {
         actualNode=null;
     }
 
-    public void addNode(Integer value){
+    public void addNode(float value){
         
         MyNode newNode= new MyNode(value);
         if(length==0){
